+++ src/main/java/com/kustacks/kuring/common/exception/InternalLogicException.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.common.error;
+package com.kustacks.kuring.common.exception;
 
 public class InternalLogicException extends BusinessException {
     public InternalLogicException(ErrorCode errorCode) {
