+++ src/main/java/com/kustacks/kuring/feedback/presentation/FeedbackControllerV1.java
@@ -1,7 +1,7 @@
 package com.kustacks.kuring.feedback.presentation;
 
-import com.kustacks.kuring.common.error.APIException;
-import com.kustacks.kuring.common.error.ErrorCode;
+import com.kustacks.kuring.common.exception.APIException;
+import com.kustacks.kuring.common.exception.ErrorCode;
 import com.kustacks.kuring.common.firebase.FirebaseService;
 import com.kustacks.kuring.feedback.business.FeedbackService;
 import com.kustacks.kuring.feedback.common.dto.SaveFeedbackV1Request;
