+++ src/main/java/com/kustacks/kuring/worker/scrap/DepartmentNoticeScraperTemplate.java
@@ -1,7 +1,7 @@
 package com.kustacks.kuring.worker.scrap;
 
-import com.kustacks.kuring.common.error.ErrorCode;
-import com.kustacks.kuring.common.error.InternalLogicException;
+import com.kustacks.kuring.common.exception.ErrorCode;
+import com.kustacks.kuring.common.exception.InternalLogicException;
 import com.kustacks.kuring.worker.scrap.deptinfo.DeptInfo;
 import com.kustacks.kuring.worker.scrap.dto.ComplexNoticeFormatDto;
 import com.kustacks.kuring.worker.scrap.dto.ScrapingResultDto;
