+++ src/main/java/com/kustacks/kuring/worker/scrap/KuisNoticeScraperTemplate.java
@@ -1,7 +1,7 @@
 package com.kustacks.kuring.worker.scrap;
 
-import com.kustacks.kuring.common.error.ErrorCode;
-import com.kustacks.kuring.common.error.InternalLogicException;
+import com.kustacks.kuring.common.exception.ErrorCode;
+import com.kustacks.kuring.common.exception.InternalLogicException;
 import com.kustacks.kuring.worker.update.notice.dto.request.KuisNoticeInfo;
 import com.kustacks.kuring.worker.update.notice.dto.response.CommonNoticeFormatDto;
 import lombok.extern.slf4j.Slf4j;
