+++ src/main/java/com/kustacks/kuring/worker/scrap/client/notice/LibraryNoticeApiClient.java
@@ -1,8 +1,8 @@
 package com.kustacks.kuring.worker.scrap.client.notice;
 
 import com.kustacks.kuring.notice.domain.CategoryName;
-import com.kustacks.kuring.common.error.ErrorCode;
-import com.kustacks.kuring.common.error.InternalLogicException;
+import com.kustacks.kuring.common.exception.ErrorCode;
+import com.kustacks.kuring.common.exception.InternalLogicException;
 import com.kustacks.kuring.common.utils.converter.DtoConverter;
 import com.kustacks.kuring.common.utils.converter.LibraryNoticeDtoToCommonFormatDtoConverter;
 import com.kustacks.kuring.worker.scrap.client.notice.property.LibraryNoticeProperties;
