+++ src/main/java/com/kustacks/kuring/worker/scrap/client/notice/RealEstateNoticeApiClient.java
@@ -1,7 +1,7 @@
 package com.kustacks.kuring.worker.scrap.client.notice;
 
-import com.kustacks.kuring.common.error.ErrorCode;
-import com.kustacks.kuring.common.error.InternalLogicException;
+import com.kustacks.kuring.common.exception.ErrorCode;
+import com.kustacks.kuring.common.exception.InternalLogicException;
 import com.kustacks.kuring.worker.scrap.client.JsoupClient;
 import com.kustacks.kuring.worker.scrap.client.notice.property.RealEstateNoticeProperties;
 import com.kustacks.kuring.worker.scrap.deptinfo.DeptInfo;
