+++ src/main/java/com/kustacks/kuring/worker/scrap/parser/staff/RealEstateStaffHtmlParser.java
@@ -1,7 +1,7 @@
 package com.kustacks.kuring.worker.scrap.parser.staff;
 
-import com.kustacks.kuring.common.error.ErrorCode;
-import com.kustacks.kuring.common.error.InternalLogicException;
+import com.kustacks.kuring.common.exception.ErrorCode;
+import com.kustacks.kuring.common.exception.InternalLogicException;
 import com.kustacks.kuring.worker.scrap.deptinfo.DeptInfo;
 import com.kustacks.kuring.worker.scrap.deptinfo.real_estate.RealEstateDept;
 import org.jsoup.nodes.Document;
