+++ src/test/java/com/kustacks/kuring/controller/FeedbackControllerTest.java
@@ -2,9 +2,9 @@ package com.kustacks.kuring.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.firebase.messaging.FirebaseMessagingException;
-import com.kustacks.kuring.common.error.APIException;
+import com.kustacks.kuring.common.exception.APIException;
 import com.kustacks.kuring.feedback.common.dto.SaveFeedbackV1Request;
-import com.kustacks.kuring.common.error.ErrorCode;
+import com.kustacks.kuring.common.exception.ErrorCode;
 import com.kustacks.kuring.feedback.presentation.FeedbackControllerV1;
 import com.kustacks.kuring.feedback.business.FeedbackService;
 import com.kustacks.kuring.common.firebase.FirebaseService;
