+++ src/test/java/com/kustacks/kuring/notice/domain/NoticeTest.java
@@ -1,6 +1,6 @@
 package com.kustacks.kuring.notice.domain;
 
-import com.kustacks.kuring.common.error.DomainLogicException;
+import com.kustacks.kuring.common.exception.DomainLogicException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
