+++ src/test/java/com/kustacks/kuring/worker/client/notice/KuisNoticeApiClientTest.java
@@ -1,7 +1,7 @@
 package com.kustacks.kuring.worker.client.notice;
 
-import com.kustacks.kuring.common.error.ErrorCode;
-import com.kustacks.kuring.common.error.InternalLogicException;
+import com.kustacks.kuring.common.exception.ErrorCode;
+import com.kustacks.kuring.common.exception.InternalLogicException;
 import com.kustacks.kuring.common.utils.converter.KuisNoticeDtoToCommonFormatDtoConverter;
 import com.kustacks.kuring.config.JsonConfig;
 import com.kustacks.kuring.worker.scrap.client.auth.ParsingKuisAuthManager;
